fix decimal double

+++ FDI.Base/Product_Size.cs
@@ -23,7 +23,7 @@ namespace FDI.Base
     
         public int ID { get; set; }
         public string Name { get; set; }
-        public Nullable<int> Value { get; set; }
+        public Nullable<double> Value { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<ProduceProductDetail> ProduceProductDetails { get; set; }
