+++ FDI.DA/DL/ProductDetailDL.cs
@@ -30,7 +30,7 @@ namespace FDI.DA.DL
                             ListProductItem = c.Shop_Product.Where(a => a.IsDelete != true).Select(v => new ProductItem
                             {
                                 ID = v.ID,
-                                PriceNew = (v.Shop_Product_Detail.Price * v.Product_Size.Value / 1000) ?? 0,
+                                PriceNew = (v.Shop_Product_Detail.Price * (decimal)v.Product_Size.Value / 1000) ?? 0,
                                 //PriceOld = v.PriceOld
                             }),
                             //ListColorProductItem = from a in c.Shop_Product
@@ -115,7 +115,7 @@ namespace FDI.DA.DL
                             ListProductItem = c.Shop_Product.Where(a => a.IsDelete != true).Select(v => new ProductItem
                             {
                                 ID = v.ID,
-                                PriceNew = (v.Shop_Product_Detail.Price * v.Product_Size.Value / 1000) ?? 0,
+                                PriceNew = (v.Shop_Product_Detail.Price * (decimal)v.Product_Size.Value / 1000) ?? 0,
                                 //PriceOld = v.PriceOld
                             }),
                             //ListColorProductItem = from a in c.Shop_Product
@@ -164,7 +164,7 @@ namespace FDI.DA.DL
                             ListProductItem = c.Shop_Product.Where(a => a.IsDelete != true).Select(v => new ProductItem
                             {
                                 ID = v.ID,
-                                PriceNew = (v.Shop_Product_Detail.Price * v.Product_Size.Value / 1000) ?? 0,
+                                PriceNew = (v.Shop_Product_Detail.Price * (decimal)v.Product_Size.Value / 1000) ?? 0,
                                 //PriceOld = v.PriceOld
                             }),
                             //ListColorProductItem = from a in c.Shop_Product
@@ -206,7 +206,7 @@ namespace FDI.DA.DL
                             ListProductItem = n.Shop_Product.Where(c => c.IsDelete != true).Select(v => new ProductItem
                             {
                                 ID = v.ID,
-                                PriceNew = (v.Shop_Product_Detail.Price * v.Product_Size.Value / 1000) ?? 0,
+                                PriceNew = (v.Shop_Product_Detail.Price * (decimal)v.Product_Size.Value / 1000) ?? 0,
                                 //PriceOld = v.PriceOld
                             }),
                             //ListColorProductItem = from a in n.Shop_Product
@@ -276,7 +276,7 @@ namespace FDI.DA.DL
                             ListProductItem = n.Shop_Product.Where(c => c.IsDelete != true).Select(v => new ProductItem
                             {
                                 ID = v.ID,
-                                PriceNew = (v.Shop_Product_Detail.Price * v.Product_Size.Value / 1000) ?? 0,
+                                PriceNew = (v.Shop_Product_Detail.Price * (decimal)v.Product_Size.Value / 1000) ?? 0,
                                 //PriceOld = v.PriceOld
                             }),
                             //ListColorProductItem = from a in n.Shop_Product
